requirements.txt
pyperclip>=1.8.2
pywin32==306
Pillow>=10.0.0 